refactor(examples): surface detail panel via Card playground page
Follow the established playground convention (SalesDynamicsWidget): extract the
channel detail panel into a reusable DetailPanelWidget component embedded at the
bottom of the Card playground page, instead of a standalone page and a separate
Examples navigation group.

- add DetailPanelWidget.vue (demo + nuxt, identical; avatar path via prop)
- embed <DetailPanelWidget> in card.vue (demo + nuxt)
- remove standalone pages/components/channel-panel.vue (demo + nuxt)
- revert the Examples navigation group in useNavigation (demo + nuxt)
