+++ playgrounds/nuxt/app/pages/components/card.vue
@@ -55,5 +55,7 @@ const attrs = reactive({
         info: 'Today vs. last 30 days'
       }"
     />
+
+    <DetailPanelWidget avatar-base="/avatar" class="mx-auto" />
   </PlaygroundPage>
 </template>
